LayoutTests/inspector/canvas/updateShader-expected.txt

Common tests for Canvas.updateShader command.


== Running test suite: Canvas.updateShader
-- Running test case: Canvas.updateShader.ProgramId.Invalid
PASS: Should produce an error.
Error: Unknown shaderType: INVALID_SHADER_TYPE
